Adding count and total
It might also be useful to tell the API client how many results are on this page and how many results there are in total. I want to show 10 results per page in the API so add a line that looks for a new property called count that's set to 10. Let's also have another property called total. That'll be the total number of results. In this case, that should be 25.

Adding Links

Finally, the API response needs to have those links! And by 'links', I mean that I want to add a new field - maybe called 'next' - whose value will be the URL to get the next page of results. Use the asserter again and change this to assertResponsePropertyExists(). Let's assert that there is an _links.next key, which means the JSON will have an _links key and a next key under that.

By moving things under _links, it makes it a little more obvious that next isn't a property of a programmer, but something different: a link.