findAllQueryBuilder() didn't exist o let's go add it!
I'll hold cmd and click to go into the ProgrammerRepository. Add the new method: public function findAllQueryBuilder(). For now, just return {dollar_sign}this->createQueryBuilder(); with an alias of programmer.

Perfect!

Creating the Pagerfanta Objects

This is all we need to use Pagerfanta. In the controller, start with {dollar_sign}adapter = new DoctrineORMAdapter() - since we're using Doctrine - and pass it the query builder. Next, create a {dollar_sign}pagerfanta variable set to new Pagerfanta() and pass it the adapter.

On the Pagerfanta object, call setMaxPerPage() and pass it 10. And then call {dollar_sign}pagerfanta->setCurrentPage() and pass it {dollar_sign}page