Using Pagerfanta to Fetch Results
Ultimately, we need Pagerfanta to return the programmers that should be showing right now based on whatever page is being requested. To get that, use {dollar_sign}pagerfanta->getCurrentPageResults(). But there's a problem: instead of returning an array of Programmer objects, this returns a type of traversable object with those programmes inside. This confuses the serializer. To fix that, create a new programmers array: {dollar_sign}programmers = [].

Next, loop over that traversable object from Pagerfanta and push each Programmer object into our simple array. This gives us a clean array of Programmer objects.

And that means we're dangerous. In createApiResponse, we still need to pass in the programmers key, but we also need to add count and total. Add the total key and set it to {dollar_sign}pagerfanta->getNbResults().

For count, that's easy: that's the current number of results that are shown on this page. Just use count({dollar_sign}programmers).

We're definitely not done, but this should be enough to return a valid response on page 1 at least. Test it out. Copy the method name and use --filter to just run that test:

bin/phpunit -c app --filter testGETProgrammersCollectionPaginated src/AppBundle/Tests/Controller/Api/ProgrammerControllerTest.php
This fails. But look closely: we do have programmers 0 through 9 in the response for page 1. It fails when trying to read the _links.next property because we haven't added those yet.