The response is returning a paginated list, and it even has extra count and total fields. Now we need to add those next, previous, first and last links. And since the response is entirely created via this PaginatedCollection class, this is simple: just add a new private {dollar_sign}_links = array(); property.
Creating and Setting the Links

To actually add links, create a new function called public function addLink() that has two arguments: the {dollar_sign}ref - that's the name of the link, like first or last - and the {dollar_sign}url. Add the link with {dollar_sign}this->_links[{dollar_sign}ref] = {dollar_sign}url; Great - now head back to the controller.

Every link will point to the same route, but with a different page query parameter. The route to this controller doesn't have a name yet, so give it one: api_programmers_collection. Copy that name and set it to a {dollar_sign}route variable.
Next, create {dollar_sign}routeParams: this will hold any wildcards that need to be passed to the route - meaning the curly brace parts in its path. This route doesn't have any, so set leave it empty. We're already setting things up to be reusable for other paginated responses.