Since pagination always looks the same, no matter what you're listing, I really want to organize my code so that pagination is effortless in the future. This took way too many lines of code.
Inside of the Pagination/ directory, create a new PHP class called PaginationFactory. There, add a new public function createCollection() method: this will create the entire final PaginatedCollection object for some collection resource. To do this, we'll need to pass it a few things, starting with the {dollar_sign}qb and the {dollar_sign}request - we'll use that to find the current page. The method will also need to know the route for the links and any {dollar_sign}routeParams it needs.

Go back to ProgrammerController, copy the logic, remove it and put it into PaginationFactory. Add the missing use statements: by auto-completing the classes DoctrineORMAdapter and Pagerfanta. Now, delete {dollar_sign}route and {dollar_sign}routeParams since those are passed as arguments. Remove the {dollar_sign}qb variable for the same reason.

In fact, move that back to ProgrammerController: we'll want it in a minute.