Next, register this as service in app/config/services.yml - let's call it pagination_factory. How creative! Set the class to PaginationFactory and pass one key for arguments: @router. Copy the service name and open ProgrammerController to hook this all up. Now, just use {dollar_sign}paginatedCollection = {dollar_sign}this->get('pagination_factory')->createCollection() and pass it the 4 arguments: {dollar_sign}qb, {dollar_sign}request, the route name - api_programmers_collection - and the route params.
Actually, most of the time you won't have route params. So head back into PaginationFactory and make that argument optional.

Much better.

Now, PhpStorm should be happy... but it's still not! It looks more like someone stole it's ice cream. Ah, I forgot to return {dollar_sign}paginatedCollection in PaginationFactory.

PhpStorm was complaining that createCollection() didn't look like it returned anything... and it was right! The robots are definitely taking over.

Run the test to see if we broke anything:

./bin/phpunit -c app --filter filterGETProgrammersCollectionPaginated src/AppBundle/Tests/Controller/Api/ProgrammerControllerTest.php
We didn't! What a delightful surprise.

Now, if you want some sweet pagination, just create a QueryBuilder, pass it into the PaginationFactory, pass that to createApiResponse and then go find some ice cream.