Don't Lose the Filter Parameter!
In the test, the URL ends in ?page=2 with no filter on it. We need to maintain the filter query parameter through our pagination. Since we have everything centralized in, PaginationFactory that's going to be easy. Add {dollar_sign}routeParams = array_merge() and merge {dollar_sign}routeParams with all of the current query parameters, which is {dollar_sign}request->query->all(). That should take care of it.

Run the tests one last time:

./bin/phpunit -c app --filter filterGETProgrammersCollectionPaginated src/AppBundle/Tests/Controller/Api/ProgrammerControllerTest.php
And we're green for filtering!