Making the URI Dynamic
This means we're almost done. To generate the real URI, we need the router. Add the __construct() method with a RouterInterface argument. I'll use the option+enter shortcut to create that property and set it.

In onPostSerialize() say {dollar_sign}programmer = {dollar_sign}event->getObject();. Because of our configuration below, we know this will only be called when the object is a Programmer. Add some inline documentation for the programmer and plug in its use statement.

Finally, for the data type {dollar_sign}this->router->generate() and pass it api_programmers_show and an array containing nickname set to {dollar_sign}programmer->getNickname().

Cool! Now, go back to services.yml and add an arguments key with just @router.

Ok, moment of truth! Run the test!
bin/phpunit -c app --filter testGETProgrammer src/AppBundle/Tests/Controller/Api/ProgrammerControllerTest.php

It's failing, but for a new reason. This time it doesn't like the app_test.php at the beginning of the link URI. Where's that coming from?

The test class extends an ApiTestCase: we made this in an earlier episode. This app already has a test environment and it configures a test database connection. If we can force every URL through app_test.php, it'll use that test database, and we'll be really happy.

We did a cool thing with Guzzle to accomplish this: automatically prefixing our requests with app_test.php. But because of that, when we generate URLs, they will also have app_test.php. That's a good thing in general, just not when we're comparing URLs in a test.