In ProgrammerControllerTest, we called this key uri.
Because, well... why not?

But when we added pagination, we included its links inside a property called _links.

That kept links separate from data. I think we should do the same thing with uri: change it to _links.self. The key self is a name used when linking to, your, 'self'

I have a different idea. Imagine we could link via annotations, like this: add @Link with 'self' inside, route = 'api_programmers_show' params = { }. This route has a nickname wildcard, so add 'nickname': and then use the expression object.getNickname().

That last part is an expression, from Symfony's expression language. You and I are going to build the system that makes this work, so I'm going to assume that we'll pass a variable called object to the expression language that is this Programmer object being serialized. Then, we just call .getNickname().

Of course, this won't work yet - in fact it'll totally bomb if you try it. But it will in a few minutes!