Creating an Annotation To create this cool system, we need to understand a bit about annotations. Every annotation - like Table or Entity from Doctrine - has a class behind it. That means we need a Link class. Create a new directory called Annotation. Inside add a new Link class in the AppBundle\Annotation namespace.
To hook this annotation into the annotations system, we need a few annotations: the first being, um, well, @Annotation. Yep, I'm being serious. The second is @Target, which will be 'CLASS'. This means that this annotation is expected to live above class declarations.

Inside the Link class, we need to add a public property for each option that can be passed to the annotation, like route and params. Add public {dollar_sign}name;, public {dollar_sign}route; and public {dollar_sign}params = array().

The first property becomes the default property, which is why we don't need to have name = 'self' when using it.

The name and route options are required, so add an extra @Required above them.

And... that's it!

Inside of Programmer, every annotation - except for the special @Annotation and @Target guys, they're core to that system - needs a use statement - we already have some for @Serializer, @Assert and @ORM. Add a use statement directly to the class itself for @Link.

This hooks the annotation up with the class we just created.