Reading the Annotation
Ok... so how do we read annotations? Great question, I have no idea. Ah, it's easy, thanks to the Doctrine annotations library that comes standard with Symfony. In fact, we already have a service available called @annotation_reader.

Inside LinkSerializationSubscriber, inject that as the second argument. It's an instance of the Reader interface from Doctrine\Common\Annotations. Call it {dollar_sign}annotationsReader.

I'll hit option+enter and select initialize fields to get that set on property.

And before I forget, in services.yml, inject that by adding @annotation_reader as the second argument.

Super easy.

Too easy, back to work! Delete all of this junk in onPostSerialize() and start with {dollar_sign}object = {dollar_sign}event->getObject(). To read the annotations off of that object, add {dollar_sign}annotations = {dollar_sign}this->annotationReader->getClassAnnotations(). Pass that a new \ReflectionObject() for {dollar_sign}object.

That's it!

Now, the class could have a lot of annotations above it, but we're only interested in the @Link annotation. We'll add an if statement to look for that in a second. But first, create {dollar_sign}links = array(): that'll be our holder for any links we find.

Now, foreach ( as ). Immediately, see if this is something we care about with if ({dollar_sign}annotation instanceof Link). At this point, the annotation options are populated on the public properties of the Link object. To get the URI, we can just say {dollar_sign}this->router->generate() and pass it {dollar_sign}annotation->route and {dollar_sign}annotation->params.

How sweet is that? Well, we're not done yet: the params contain an expression string... which we're not parsing yet. We'll get back to that in a second.

Finish this off with {dollar_sign}links[{dollar_sign}annotation->name] = {dollar_sign}uri;. At the bottom, finish with the familiar {dollar_sign}visitor->addData() with _links set to {dollar_sign}links;. Other than evaluating the expression, that's all the code you need.

Check this out by going to /api/programmers in the browser. Look at that! The embedded programmer entities actually have a link called self. It worked!

Of course, the link is totally wrong because we're not evaluating the expression yet. But, we're really close.