Before we fix the expression stuff, remove the class option from getSubscribedEvents() because we want this to be called for all classes.
Avoiding Duplicated _links

When you do that, things still work. But now: clear your cache in the terminal:

./app/console cache:clear
That's not normally something you need to do - but the JMSSerializerBundle doesn't properly update its cache when you change this option. Refresh again.

Ah, but wait: one of the things we're serializing is the paginated collection itself, which already has a _links property.