Evaluating the Expression
Refresh the browser again. Things look good! Time to evaluate the expression.

To use the expression language, we just need to create an ExpressionLanguage object. We could register this as a service, but I'll take a shortcut and instantiate a new expression language right inside the constructor: {dollar_sign}this->expressionLanguage = new ExpressionLanguage().

That class lives in the ExpressionLanguage component.

Rename that property to expressionLanguage. Later, if I do want to register this as a service instead of creating it new right here, that'll be really easy.

Wrap {dollar_sign}annotation->params in a call to {dollar_sign}this->resolveParams() and pass it the params and the object, since we'll need to pass that into the expression itself.

Add the new private function resolveParams() and then loop over : foreach ({dollar_sign}params as {dollar_sign}key => {dollar_sign}param).

For each param, we'll replace it with {dollar_sign}this->expressionLanguage->evaluate(). Pass it {dollar_sign}param - that's the expression. Next, since the expressions are expecting a variable called object, pass an array as the second argument with an object key set to {dollar_sign}object. And let's not forget our  argument to this method!

Finally, wrap this up return {dollar_sign}params;. Now, each parameter is evaluated through the expression language, which is a lot like Twig.

Ok, back to the browser. There it is! How about our test?

./bin/phpunit -c app --filter testGETProgrammer src/AppBundle/Tests/Controller/Api/ProgrammerControllerTest.php
Hey, they're passing too! Amazing!

In just a few short minutes, we made an entirely reusable linking system. I will admit that this idea was stolen from a library called Hateoas(https://github.com/willdurand/Hateoas). Now, don't you feel dangerous?