Testing the Deep Functionality
Let's start with a quick test: copy part of testGETProgramer() and name the new method testGETProgrammerDeep(). Now, add a query parameter called ?deep.

The idea is simple: if the client adds ?deep=1, then the API should expose more embedded objects. Use the asserter to say assertResponsePropertyExists(), pass that the  and the property we'll expect, which is user. Since this will be an entire user object, check specifically for user.username.

Serialization Groups
If you look at this response in the browser, we definitely do not have a user field. But there are only two little things we need to do to add it.