Add @Serializer\Groups() and use a new group called deep to avoid user property to be always included.
Here's the idea: when you serialize, each property belongs to one or more 'groups'. If you don't include the @Serializer\Groups annotation above a property, then it will live in a group called Default - with a capital D. Normally, the serializer serializes all properties, regardless of their group. But you can also tell it to serialize only the properties in a different group, or even in a set of groups. We can use groups to serialize the user property under only certain conditions.